Clean up update pipeline prior to logging
