+++ src/MongoDB.EntityFrameworkCore/Serializers/SerializationHelper.cs
@@ -58,57 +58,6 @@ internal static class SerializationHelper
         throw new InvalidOperationException($"Document element '{elementName}' is missing but required.");
     }
 
-    internal static void WriteKeyProperties(IBsonWriter writer, IUpdateEntry entry)
-    {
-        var keyProperties = entry.EntityType.FindPrimaryKey()
-            .Properties
-            .Where(p => !p.IsShadowProperty() && p.GetElementName() != "").ToArray();
-
-        if (!keyProperties.Any()) return;
-
-        bool compoundKey = keyProperties.Length > 1;
-        if (compoundKey)
-        {
-            writer.WriteName("_id");
-            writer.WriteStartDocument();
-        }
-
-        foreach (var property in keyProperties)
-        {
-            object? propertyValue = entry.GetCurrentValue(property);
-            var serializationInfo = GetPropertySerializationInfo(property);
-            string? elementName = serializationInfo.ElementPath?.Last() ?? serializationInfo.ElementName;
-            WriteProperty(writer, elementName, propertyValue, serializationInfo.Serializer);
-        }
-
-        if (compoundKey)
-        {
-            writer.WriteEndDocument();
-        }
-    }
-
-    internal static void WriteNonKeyProperties(IBsonWriter writer, IUpdateEntry entry, Func<IProperty, bool>? propertyFilter = null)
-    {
-        var properties = entry.EntityType.GetProperties()
-            .Where(p => !p.IsShadowProperty() && !p.IsPrimaryKey() && p.GetElementName() != "")
-            .Where(p => propertyFilter == null || propertyFilter(p))
-            .ToArray();
-
-        foreach (var property in properties)
-        {
-            var propertyValue = entry.GetCurrentValue(property);
-            var serializationInfo = GetPropertySerializationInfo(property);
-            WriteProperty(writer, serializationInfo.ElementName, propertyValue, serializationInfo.Serializer);
-        }
-    }
-
-    private static void WriteProperty(IBsonWriter writer, string elementName, object value, IBsonSerializer serializer)
-    {
-        writer.WriteName(elementName);
-        var context = BsonSerializationContext.CreateRoot(writer);
-        serializer.Serialize(context, value);
-    }
-
     internal static BsonSerializationInfo GetPropertySerializationInfo(IReadOnlyProperty property)
     {
         var serializer = CreateTypeSerializer(property.ClrType, property);
