+++ src/MongoDB.EntityFrameworkCore/Storage/IMongoClientWrapper.cs
@@ -15,24 +15,29 @@
 
 using System;
 using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
 using MongoDB.Driver;
 using MongoDB.EntityFrameworkCore.Query;
 
 namespace MongoDB.EntityFrameworkCore.Storage;
 
 /// <summary>
+/// For internal use only. Interface may change between minor versions.
 /// Provides the interface between the MongoDB Entity Framework provider
 /// and the underlying <see cref="IMongoClient"/>.
 /// </summary>
 public interface IMongoClientWrapper
 {
     /// <summary>
-    /// The <see cref="IMongoDatabase"/> this client is connected to.
+    /// Get an <see cref="IMongoCollection{T}"/> for the given <paramref name="collectionName"/>;
     /// </summary>
-    public IMongoDatabase Database { get; }
+    /// <param name="collectionName">The name of the collection to get a collection for.</param>
+    /// <typeparam name="T">The type of data that will be returned by the collection.</typeparam>
+    /// <returns>The <see cref="IMongoCollection{T}"/> </returns>
+    public IMongoCollection<T> GetCollection<T>(string collectionName);
 
     /// <summary>
-    /// For internal use only. Interface may change between minor versions.
     /// A query and the associated metadata and provider needed to execute that query.
     /// </summary>
     /// <param name="executableQuery">The <see cref="MongoExecutableQuery"/> that will be executed.</param>
@@ -41,5 +46,18 @@ public interface IMongoClientWrapper
     /// <returns>An <see cref="IEnumerable{T}"/> containing the results of the query.</returns>
     public IEnumerable<T> Execute<T>(MongoExecutableQuery executableQuery, out Action log);
 
-    // TODO: Add item update/delete/insert operations
+    /// <summary>
+    /// Save updates to a MongoDB database.
+    /// </summary>
+    /// <param name="updates">The updates to save to the database.</param>
+    /// <returns>The number of affected documents.</returns>
+    public long SaveUpdates(IEnumerable<MongoUpdate> updates);
+
+    /// <summary>
+    /// Save updates to a MongoDB database asynchronously.
+    /// </summary>
+    /// <param name="updates">The updates to save to the database.</param>
+    /// <param name="cancellationToken"></param>
+    /// <returns>A task that when completed gives the number of affected documents.</returns>
+    public Task<long> SaveUpdatesAsync(IEnumerable<MongoUpdate> updates, CancellationToken cancellationToken);
 }
